refactor(acp): inline SaaS versioned secret persistence
